Reorganize namespace
