+++ src/main/php/PHPMD/Cache/ResultCacheEngineFactory.php
@@ -2,6 +2,7 @@
 
 namespace PHPMD\Cache;
 
+use PHPMD\Cache\Model\ResultCacheConfig;
 use PHPMD\TextUI\CommandLineOptions;
 
 class ResultCacheEngineFactory
