+++ src/main/php/PHPMD/Cache/ResultCacheStateFactory.php
@@ -2,6 +2,8 @@
 
 namespace PHPMD\Cache;
 
+use PHPMD\Cache\Model\ResultCacheState;
+
 class ResultCacheStateFactory
 {
     /**
