+++ src/main/php/PHPMD/Cache/ResultCacheWriter.php
@@ -2,6 +2,8 @@
 
 namespace PHPMD\Cache;
 
+use PHPMD\Cache\Model\ResultCacheState;
+
 class ResultCacheWriter
 {
     /** @var string */
